+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ignore `About.tips` in windows
 - allow `/` to trigger path complete for windows
 - more default alias for `Fs` operation
+- skip invalid escape like `\windows` in windows
 
 ## [0.7.5]
 - update rustyline to 17.0.